+++ src/api.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once j(LIBS_DIR, 'request-handler.php');
+require_once j(ENDPOINTS_DIR, 'screenshot.php');
 
 /**
  * Dispatch the API request
@@ -9,5 +10,15 @@ require_once j(LIBS_DIR, 'request-handler.php');
  */
 function dispatch(): void
 {
-    echo 'Hello World!';
+    $url = get_value('url');
+
+    if (!$url) {
+        echo "No \"url\" param was given.\n";
+        return;
+    } else if (!filter_var($url, FILTER_VALIDATE_URL)) {
+        echo "No valid url was given. Make sure the required HTTP protocol is there\n";
+        return;
+    }
+
+    process_screenshot($url);
 }