+++ CHANGELOG.md
@@ -15,3 +15,4 @@ The project history of changes.
 - Implement basic screenshotting scipt, functional. First with chrome
 - Prepare python execution enviroment
 - Implement debugging and logging systems
+- Create screenshot basic script call